Fix: SimDynamics was not actuating joints properly when bullet was compiled with floats Minor changes: added convenient parameter settings for some methods (collision detection) Minor Fix: added some explicit casts to avoid double/float warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/simox/code/trunk@606 042f3d55-54a8-47e9-b7fb-15903f145c44

+++ SimDynamics/DynamicsEngine/BulletEngine/BulletEngine.cpp
@@ -395,7 +395,7 @@ void BulletEngine::updateRobots(btScalar timeStep)
 {
 	for (size_t i=0; i < robots.size();i++)
 	{
-		robots[i]->actuateJoints(timeStep);
+		robots[i]->actuateJoints(double(timeStep));
 		robots[i]->updateSensors();
 	}
 }
